requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# Data Science and Machine Learning
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
joblib>=1.1.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# API Framework
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0

# Additional utilities
python-multipart>=0.0.6
